Add checking in most cases with flash errors. And many tweaks...
This commit adds different colour alerts, and a template
for flash error message which is added in most pages.

We also remove Gzip which is broken in panic'd pages.

We validate some fields with RegExp and by converting
to markdown, back to plaintext (removing HTML tags) and
measuring the length.

We also use session to pass over the text in case that the
post or comment has some issues, so content is not lost.